Renamed Filter to ResourceMenu in all possible locations
